worker/matcher.ts: renames file.
